requirements.txt
# ✅ FastAPI & Web Server
fastapi
uvicorn

# ✅ MongoDB Database
motor

# ✅ User Authentication (Password Hashing & JWT)
passlib[bcrypt]
pyjwt

# ✅ Machine Learning (Glucose Prediction)
numpy
pandas
scikit-learn
xgboost
joblib

# ✅ OCR (Medical Report Extraction)
pytesseract
pdf2image
pillow

# ✅ Image Processing
opencv-python-headless  # Headless version for Docker compatibility

# ✅ PDF Processing
reportlab

# ✅ Firebase Cloud Messaging (Push Notifications)
firebase-admin

# ✅ Other Utilities
python-multipart  # Required for file uploads in FastAPI
requests  # For making HTTP requests if needed
